memory-mcp-server/pyproject.toml
[project]
name = "memory-mcp-server"
version = "0.1.0"
description = "MCP server for memory storage with full-text search"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.12.3",
    "psycopg[binary]>=3.2.10",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
]

[project.scripts]
memory-server = "src.mcp_server:main"
